refactor: input view 로직 합쳐서 interface 구현하기
