--- src/main/java/lotto/view/input/Input.java
@@ -1,16 +1,0 @@
-package lotto.view.input;
-
-import camp.nextstep.edu.missionutils.Console;
-
-public interface Input {
-
-    static String readLine(){
-        return Console.readLine();
-    }
-
-    default void printErrorMessage(RuntimeException e) {
-        System.out.println(e.getMessage());
-        System.out.println();
-    }
-
-}
